updated nickname calls on reviews cards, also refactored review card code to get ride of an extra component that wasn't needed
