+++ Front_end/src/Components/ReviewsPage/ReviewCard.js
@@ -1,27 +1,41 @@
 import React, { Component, Fragment } from "react";
 import styled from "styled-components";
 import { Container, Row, Col, Jumbotron, CardImg, Button } from "reactstrap";
-import ViewStars from "../StarsRating/ViewStars";
 import axios from "axios";
+import ViewStars from "../StarsRating/ViewStars";
 import { instanceOf } from "prop-types";
 import { withCookies, Cookies } from "react-cookie";
 
-class TrackReviewCard extends Component {
+class AlbumReviewCard extends Component {
   static propTypes = {
     cookies: instanceOf(Cookies).isRequired
   };
   constructor(props) {
     super(props);
     this.state = {
+      users: [],
       nickname: ""
     };
   }
 
+  getUser() {
+    axios
+      .get("https://labs9-car-reviews.herokuapp.com/users")
+      .then(response => {
+        const users = response.data;
+        const newState = Object.assign({}, this.state, {
+          users: users
+        });
+        this.setState(newState);
+      })
+      .catch(err => console.log(err));
+  }
+
   getNickname(userID) {
     axios
-      .get(`https://labs9-car-reviews.herokuapp.com/user/${userID}/nickname`)
+      .get(`https://labs9-car-reviews.herokuapp.com/users/${userID}/nickname`)
       .then(response => {
-        const userNickname = response.data;
+        const userNickname = response.data[0].nickname;
         const newState = Object.assign({}, this.state, {
           nickname: userNickname
         });
@@ -31,8 +45,10 @@ class TrackReviewCard extends Component {
   }
 
   componentDidMount() {
-    this.getNickname(this.state.nickname);
+    this.getUser();
+    this.getNickname(this.props.review.userID);
   }
+
   render() {
     console.log(this.props.review);
     return (
@@ -53,16 +69,21 @@ class TrackReviewCard extends Component {
                   <img
                     src={require("../../Images/RecordThumb.png")}
                     alt="Default profile image"
-                    style={{ maxWidth: '150px'}}
+                    style={{ maxWidth: "150px" }}
                   />
                 </div>
-                <div><strong>Nickname: </strong>{this.state.nickname}</div>
+                <div>
+                  <strong>Nickname: </strong>
+                  {this.state.nickname}
+                </div>
                 <div>Member status</div>
-                <div><strong>Reviews: </strong></div>
+                <div>
+                  <strong>Reviews: </strong>
+                </div>
               </Col>
               <Col md="8" style={{ padding: "1rem 5rem" }}>
                 <Row style={{ display: "flex" }}>
-                <ViewStars rating={this.props.review.rating} />
+                  <ViewStars rating={this.props.review.rating} />
                   <p style={{ margin: "auto" }}>Date Written: (DATE)</p>
                 </Row>
                 <Row>
@@ -79,4 +100,4 @@ class TrackReviewCard extends Component {
   }
 }
 
-export default withCookies(TrackReviewCard);
+export default withCookies(AlbumReviewCard);
